OCR/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.0
gunicorn==22.0.0
imageio==2.34.0
itsdangerous==2.2.0
Jinja2==3.1.3
lazy_loader==0.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pytweening==1.2.0
scikit-image==0.23.2
scipy==1.13.0
tifffile==2024.4.18
Werkzeug==3.0.2
